remove unused variables / release notes
Typography:

First thing you'll note is the use of the typography I had proposed last
week. I do think it has presence and strikes a nice balance between
friendliness and seriousness. So this our current point-of-reference
(meaning that it's possible that we can change, but we use this unless we
find something better).

Also note the color scheme. I went a bit darker than I started with (and
considering your last note about colors) because once I started to get a
lot of content on the home page, which feels busy to me at the moment,
perhaps a desired effect, the darker color was less of a distraction.

With colors as with so many things, there are design reasons for my
choices, but they're also subjective, so please do let me know what you
think.

Of course, everything here is subject to change and refinement (just
reinforcing this concept! :) ).

Carousel:

Next, you'll note the new carousel. Carousel's have to be used judiciously
because they often don't get paid attention to beyond the first slide, and
the represent a performance overhead (i.e. it takes more time to load
scripts to make them work, potentially slowing down page loads).

What I've done is to create a carousel with the small, card-like blocks.
Doing so allows us to get quite a bit of information in (on a desktop you
can easily take in several slides) and makes the carousel lightweight as
well as not overwhelming the page.

You'll note that the carousel has both images (the map, and Elvis Costello)
and solid color backgrounds. I wanted to make it so that there was no onus
on your team to HAVE to come up with images. So, you can just put some text
on there and it will show up on a nice colored card (we can refine the
colors too).

So this will allow you to highlight sales, special events, basically
anything you like, and we'll always keep the map as the first slide.
Hopefully that's good for the map (i.e. a good way to accomplish having the
map front-and-center without distracting from other things, but we'll see).

Home page body:

Below that, you'll see news posts on the left and events on the right. News
posts are in a card-view, and you'll see here too I've made it so you're
not required to use an image (see the "It's Nothing against the Book you're
Buying" for an example of what the post looks like when there's no featured
image). I still have to style the article pages.

The events listings on the home page are kept to the bare minimum so that
one can take in as many events as possible at a glance. I've been thinking
of a lot of different ways to display events, but I think to start off with
we keep it simple. Of course, let me know if you feel we need more info
there.

Note that I have events, and then I've segregated kids events in their own
tab. I think, of the various types of events you offer, kids events are one
that someone might be specifically looking for, hence that placement. I
don't think it's a good idea to add a bunch of choices there (i.e. break
them all up), but I do think showing kids events is important for a segment
of your audience.

Also note that on smaller mobile devices, the event listing will show above
the news posts (same goes for the right-side content on interior pages:
they'll show on the top for mobile), that's just a bit tricky to code and I
haven't done it yet.

The home page is quite busy, particularly with the carousel. The "designer"
me says pull back and make things as clean as possible, but the "business"
me says show a lot of activity as long as we get the most important
utilitarian stuff in so people don't miss them.

Events page:

The events page itself is still evolving design-wise (and, note not all
book covers are in on all events, I'll get them in, but see below too), but
again, starting with just simple listing.

Programs page:

We discussed this before, but addressing here because it's the first time
you've seen my conception.

The goal with the page is to show each type of event (with emphasis on the
non-main events) to show that you offer those specific things and to list
upcoming events for those categories with them. So, for instance, if
someone's interested in the Small Press book club, she can see the upcoming
dates (you could also be listing the next book to be discussed on these).

What's missing:

I have to put in some book covers on events. I've been working on tapping
into an API instead of adding books directly to the site. So far, little
luck here. There are a lot of these databases out there, but with spotty
quality and information.

We need to determine the content of the giftsmith page. From what I can
tell most of these sorts of landing pages (specials page, giftsmith, etc)
will be listings with images, copy, etc).

I haven't finished the search yet. Right now it's only on the mobile menu
(and doesn't do anything; for the moment it's there to show you), but I'll
be incorporating it soon. Of course, at this point there's not much to
search for.

So at this point I think it's good for you all to have a look and see how
you feel about the direction we're taking. I've tried to make things quite
polished for this stage of design (though to me it's still rough), but the
overall goal is to get a public site up as soon as we can so we and your
users can interact with it in the real-world and let it evolve. We tend to
go without formal announcements, but on the site we can encourage people to
contact us with comments about using the site - that sort of thing can be
invaluable and your customers can tell us a lot more than we can all make
up sitting in a room.

I'll address your First Editions club in a separate email.

Here's the link: http://brooklinebooksmith-151013-114847.surge.sh

Signed-off-by: budparr <budparr@sonnetmedia.net>